fixed bug for finding first visible design -- needed to order linked list of design types first
